bot-trading/requirements.txt
ib_insync>=0.9.86
pandas>=2.0.0
numpy>=1.24.0
aiohttp>=3.9.0
pandas_market_calendars>=4.3.0
python-dotenv>=1.0.0
pytz>=2024.1
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-timeout>=2.2.0
yfinance>=0.2.0
